+++ tests/unit/test_plotting/test_quick_plot.py
@@ -7,9 +7,7 @@ from tempfile import TemporaryDirectory
 
 
 class TestQuickPlot:
-    _solver_args = [pybamm.CasadiSolver()]
-    if pybamm.has_idaklu():
-        _solver_args.append(pybamm.IDAKLUSolver())
+    _solver_args = [pybamm.CasadiSolver(), pybamm.IDAKLUSolver()]
 
     @pytest.mark.parametrize("solver", _solver_args)
     def test_simple_ode_model(self, solver):